backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6
Pillow>=10.0.0
pytesseract>=0.3.10
requests>=2.31.0
pydantic>=2.0.0
